+++ src/components/CategoryChips/CategoryChips.jsx
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { useEffect, useState } from "react";
 import "../CategoryChips/CategoryChips.css";
-import { useData } from "../../context/dataContext/dataContext";
+import { useData } from "../../context";
 import { useToast } from "../../custom-hooks/useToast";
 
 const CategoryChips = () => {
