+++ src/context/index.js
@@ -1,0 +1,9 @@
+export {
+  useAuth,
+  AuthProvider,
+} from "../context/authContext/authenticationContext.jsx";
+export { useData, DataProvider } from "../context/dataContext/dataContext.jsx";
+export {
+  useServices,
+  ServiceProvider,
+} from "../context/servicesContext/servicesContext.jsx";
