+++ src/pages/Athentication/Login/Login.jsx
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { useEffect } from "react/cjs/react.production.min";
-import { useAuth } from "../../../context/authContext/authenticationContext";
+import { useAuth } from "../../../context";
 
 const Login = () => {
   const initialFormDetails = {
