+++ src/pages/WatchLaterPage/WatchLaterPage.jsx
@@ -1,5 +1,5 @@
 import { VideoCard } from "../../components";
-import { useServices } from "../../context/servicesContext/servicesContext";
+import { useServices } from "../../context";
 
 const WatchLaterPage = () => {
   const { state } = useServices();
